agents/video_validator/requirements.txt
transformers>=4.40.0
torch>=2.0.0
Pillow
requests
python-dotenv
einops
timm
accelerate
